Acceleration Profile Data Added
Added data that provides x, y, and z components of acceleration of the solar lantern during its typical movement states - stationary and mobile. Sequence number is added for each data point. The test was carried out for around 39 minutes, with the lantern kept stationary for 3 minutes, and mobile for another 3 minutes.